+++ R/DataSummary.R
@@ -1,13 +1,16 @@
 #' DataSummary
 #'
 #' @description function gives summary of the dataset.
-#' @usage DataSummary(data,missing=list(NA),wt=NULL,entropy_method="emp",sparkline=FALSE)
+#' @usage DataSummary(data,missing=list(NA),wt=NULL,entropy=FALSE,entropy_method="emp",sparkline=FALSE)
 #' @param data This could be data frame or a vector.
 #' @param missing list of possible values stand for missing.
-#' @param wt For a sampled dataset, you may want to specify the wieght for stats calcualted.  It can be a character which is a column name in the dataset provided, 
-#' or integer (numeric) weights vector.
-#' @param entropy_method The name of the entropy estimator. The `infotheo` package implements four estimators : "emp", "mm", "shrink", "sg" (default:"emp").
-#' @param sparkline logical. If true, a string of level percentage will be generated, which can be used later in `shiny` app with `sparkline` package.
+#' @param wt For a sampled dataset, you may want to specify the wieght for stats calcualted.  
+#' It can be a character which is a column name in the dataset provided, or integer (numeric) weights vector.
+#' @param entropy logical. whether to include entropy as part of output. It will take quite a lot of time.
+#' @param entropy_method The name of the entropy estimator. 
+#' The `infotheo` package implements four estimators : "emp", "mm", "shrink", "sg" (default:"emp").
+#' @param sparkline logical. If true, a string of level percentage will be generated, 
+#' which can be used later in `shiny` app with `sparkline` package.
 #' @details This function provides a data summary including min, max, number of unique values and number if missing values.
 #' The min and max will ignore missing value in the data.  The input should be a `data.frame`.
 #' @author Sixiang Hu
@@ -18,21 +21,23 @@
 #' @examples
 #' DataSummary(mtcars,missing=list(NA,".","Unknown",-1))
 
-DataSummary <- function(data,missing=list(NA),wt=NULL,entropy_method="emp",sparkline=FALSE){
+DataSummary <- function(data,missing=list(NA),wt=NULL,entropy=FALSE,entropy_method="emp",sparkline=FALSE){
   UseMethod("DataSummary",data)
 }
 
 #' @export
 #' @rdname DataSummary
-DataSummary.data.frame <- function(data,missing=list(NA),wt=NULL,entropy_method="emp",sparkline=FALSE){
+DataSummary.data.frame <- function(data,missing=list(NA),wt=NULL,
+                                   entropy=FALSE,entropy_method="emp",sparkline=FALSE){
   setDT(data)
-  DataSummary.data.table(data,missing,wt,entropy_method,sparkline)
+  DataSummary.data.table(data,missing,wt,entropy,entropy_method,sparkline)
 }
 
 #' @export
 #' @rdname DataSummary
 #' 
-DataSummary.data.table <- function(data,missing=list(NA),wt=NULL,entropy_method="emp",sparkline=FALSE){
+DataSummary.data.table <- function(data,missing=list(NA),wt=NULL,
+                                   entropy=FALSE,entropy_method="emp",sparkline=FALSE){
   
   if (is.null(wt)) weight <- rep(1,nrow(data))
   else if ( is.character(wt) ) weight <- data[[wt]]
@@ -73,7 +78,8 @@ DataSummary.data.table <- function(data,missing=list(NA),wt=NULL,entropy_method=
     }
   })]
   
-  dsEntrophy <- data[,sapply(data,function(x) infotheo::entropy(infotheo::discretize(x), method = entropy_method))]
+  dsEntropy <- rep(0,length(dsName))
+  if(entropy) dsEntropy <- data[,sapply(data,function(x) infotheo::entropy(infotheo::discretize(x), method = entropy_method))]
 
   if (sparkline) {
     dsStr <- data[,sapply(data,function(x) {
@@ -90,7 +96,7 @@ DataSummary.data.table <- function(data,missing=list(NA),wt=NULL,entropy_method=
                       "Mean"=dsMean,
                       "Min"=dsMin,
                       "Max"=dsMax,
-                      "Entrophy"=round(dsEntrophy,6),
+                      "Entropy"=round(dsEntropy,6),
                       "%Distribution"=dsStr,
                       stringsAsFactors = FALSE))
   }else{
@@ -102,7 +108,7 @@ DataSummary.data.table <- function(data,missing=list(NA),wt=NULL,entropy_method=
                       "Mean"=dsMean,
                       "Min"=dsMin,
                       "Max"=dsMax,
-                      "Entrophy"=round(dsEntrophy,6),
+                      "Entropy"=round(dsEntropy,6),
                       stringsAsFactors = FALSE))
   }
 }